add routing and slight refactoring of app layout
