+++ HeroesRandomiser.Web/ClientApp/src/App.jsx
@@ -1,6 +1,10 @@
 import React, { Component } from 'react';
 import { Provider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
+
 import 'normalize.css';
+import { ThemeProvider } from 'styled-components';
+import theme from './style/_theme';
 
 import configureStore from './store/configureStore';
 import Layout from './containers/Layout';
@@ -11,7 +15,11 @@ class App extends Component {
   render() {
     return (
       <Provider store={store}>
-        <Layout />
+        <ThemeProvider theme={theme}>
+          <BrowserRouter>
+            <Layout />
+          </BrowserRouter>
+        </ThemeProvider>
       </Provider>
     );
   }
