+++ HeroesRandomiser.Web/ClientApp/src/containers/AboutPage.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class AboutPage extends Component {
+    render() {
+        return (
+            <section>
+                <p>About</p>
+            </section>
+        );
+    }
+}
+
+export default AboutPage;
