+++ HeroesRandomiser.Web/ClientApp/src/style/pageLayout.js
@@ -1,8 +1,9 @@
 import styled from 'styled-components';
+import { withRouter } from 'react-router-dom';
 
 import { media } from './_mediaTemplates';
 
-export const Viewport = styled.div`
+export const Viewport = withRouter(styled.div`
     display: flex;
     flex-direction: column;
     width: 100%;
@@ -10,9 +11,9 @@ export const Viewport = styled.div`
     background-image: url(https://prismic-io.s3.amazonaws.com/heroesrandomiser%2Ffac1e1c3-8bf2-467b-b3a2-bdbc5140206e_gabriel-garcia-marengo-6132-unsplash-shrunk.jpg);
     background-size: cover;
     background-position: center;   
-`;
+`);
 
-export const MainBody = styled.section`
+export const MainBody = withRouter(styled.main`
     flex: 1;
     margin: auto;
     min-height: 500px;
@@ -24,4 +25,4 @@ export const MainBody = styled.section`
     ${media.desktop`
         width: 60% 
     `}
-`;
+`);
